requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
flask-bcrypt==1.0.1
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu>=1.8.0
gunicorn==21.2.0

